backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
psycopg2-binary==2.9.9
python-dotenv==0.19.0
python-dateutil==2.8.2
sendgrid==6.9.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
gunicorn==20.1.0
python-slugify==5.0.2
PyJWT==2.3.0
requests==2.26.0
Werkzeug==2.0.1
email-validator==1.1.3
